WIP Finding path within grid map
Create a UI using QML that consists of a grid of boxes.
Each box can be selected to be either an obstacle, a
begin/end location, or unobstructed. Create two buttons,
one for resetting the map, and another for calculating the
path between the begin/end locations.

The state of the grid is passed to the C++ backend for
determing the path, and then the path is sent back to the
UI to be displayed.

To do:
- There is an issue with the path determination. To replicate:
  - Set the begin/end locations to coordinates (0, 3) and
  (3, 4)
  - Obstruct coordinates (0, 4), (1, 3), (2, 2), (3, 1)
  - The path does not reach the end location.