Refactor QML to separate functionality
Separate functionality of the occupancy grid with
the vertexes of the grid. The goal is to make it
easier to switch to a different type of vertex,
without needing to heavily refactor GridSelectBox

The Vertex is dependent on GridSelectBox, but
GridSelectBox is not dependent on the vertex type, as
long as the type has the required signals it can
connect to.

Create the RectangleVertex QML type that encapsulates
a rectangular vertex.

Update/fix doxygen documentation for C++ and QML files.

Update README with information on how to create doxygen
documentation.